Adding more data to the Events model
