+++ nps-kiosk-app/src/app/event-page/event-page.component.html
@@ -44,5 +44,21 @@
         </div>
       </div>
     </div>
+    <div class="row">
+      <div class="col">
+        <ul class="list-group" *ngFor="let displayElement of receivedObject.getDisplayElements()">
+          <li class="list-group-item card my-2 mx-1"
+              *ngIf="displayElement.getDisplayElementType() === DISPLAY_IMAGE">
+            <img [src]="displayElement.getUrl()"
+                 [alt]="displayElement.getDescription()" style="width:50%;"
+                 class="float-right">
+            <div>
+              <p class="font-weight-bold mt-1 mb-0">{{ displayElement.getTitle() }}</p>
+              <p class="my-1" [innerHTML]="displayElement.getDescription() | keepHtml"></p>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </div>
   </div>
 </section>
